+++ tests/run_instruction_checks.sh
@@ -10,5 +10,5 @@ set -e
 ./build/tests/c/instruction_checks/r3000gte_disasm.elf
 ./build/tests/c/instruction_checks/r4000allegrex_disasm.elf
 ./build/tests/c/instruction_checks/r4000allegrex_vfpu_disasm.elf
+./build/tests/c/instruction_checks/r5900_disasm.elf
 ./build/tests/c/instruction_checks/r5900_trunc_cvt.elf
-./build/tests/c/instruction_checks/r5900_vcallms.elf